+++ scripts/pylint_custom_plugin/tests/test_files/enum_checker_acceptable.py
@@ -1,0 +1,8 @@
+# Test file for enum checker
+from enum import Enum
+from six import with_metaclass
+from azure.core import CaseInsensitiveEnumMeta
+
+class EnumPython2(with_metaclass(CaseInsensitiveEnumMeta, str, Enum)):
+    ONE = "one"
+    TWO = "two"