+++ sdk/cosmos/azure-cosmos/azure/cosmos/cosmos_client.py
@@ -23,7 +23,7 @@
 """
 
 from typing import Any, Dict, Optional, Union, cast, Iterable, List  # pylint: disable=unused-import
-
+import warnings
 from azure.core.tracing.decorator import distributed_trace  # type: ignore
 
 from ._cosmos_client_connection import CosmosClientConnection
@@ -97,15 +97,23 @@ def _build_connection_policy(kwargs):
         policy.SSLConfiguration = ssl
 
     # Retry config
-    retry = kwargs.pop('retry_options', None) or policy.RetryOptions
+    retry_options = kwargs.pop('retry_options', None)
+    if retry_options is not None:
+        warnings.warn("This option has been deprecated and will be removed from the SDK in a future release.",
+                      DeprecationWarning)
+    retry_options = policy.RetryOptions
     total_retries = kwargs.pop('retry_total', None)
-    retry._max_retry_attempt_count = total_retries or retry._max_retry_attempt_count
-    retry._fixed_retry_interval_in_milliseconds = kwargs.pop('retry_fixed_interval', None) or \
-        retry._fixed_retry_interval_in_milliseconds
+    retry_options._max_retry_attempt_count = total_retries or retry_options._max_retry_attempt_count
+    retry_options._fixed_retry_interval_in_milliseconds = kwargs.pop('retry_fixed_interval', None) or \
+        retry_options._fixed_retry_interval_in_milliseconds
     max_backoff = kwargs.pop('retry_backoff_max', None)
-    retry._max_wait_time_in_seconds = max_backoff or retry._max_wait_time_in_seconds
-    policy.RetryOptions = retry
-    connection_retry = kwargs.pop('connection_retry_policy', None) or policy.ConnectionRetryConfiguration
+    retry_options._max_wait_time_in_seconds = max_backoff or retry_options._max_wait_time_in_seconds
+    policy.RetryOptions = retry_options
+    connection_retry = kwargs.pop('connection_retry_policy', None)
+    if connection_retry is not None:
+        warnings.warn("This option has been deprecated and will be removed from the SDK in a future release.",
+                      DeprecationWarning)
+    connection_retry = policy.ConnectionRetryConfiguration
     if not connection_retry:
         connection_retry = ConnectionRetryPolicy(
             retry_total=total_retries,