+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -5,6 +5,10 @@
 #### Breaking Changes
 - Method signatures have been updated to use keyword arguments instead of positional arguments for most method options in the async client.
 
+#### Other Changes
+ - Added samples showing how to configure retry options for both the sync and async clients.
+ - Deprecated the `connection_retry_policy` and `retry_options` options in the sync client.
+
 ### 4.3.0b4 (2022-04-07)
 
 #### Features Added